Build the generated templates from templates
Added options to the grid for pulling all markup for generating header
and row templates in from smaller templates.

Modified generateHeaderTemplate and generateRowTemplate to utilize these
templates.  Compositing is now done using DOM fragments with jQuery
instead of with a string builder....

Hmmm, that's leaving a jquery dependency in there that the original
KoGrid doesn't want, I think.  Oh well, fix it later.